Using click commands for both web app & commands.
